+++ classes/local/restoreactions/uninstall_missing_plugins.php
@@ -42,7 +42,7 @@ class uninstall_missing_plugins extends restore_action {
             return;
         }
 
-        $needsupgrade = recalc_version_hash::fetch_core_version() != (float)$CFG->version;
+        $needsupgrade = recalc_version_hash::fetch_core_version()['version'] != (float)$CFG->version;
         \core_plugin_manager::reset_caches();
         $pluginman = \core_plugin_manager::instance();
         /** @var \core\plugininfo\base[][] $plugininfo */
@@ -91,6 +91,6 @@ class uninstall_missing_plugins extends restore_action {
 
         $logger->add_to_log('...done');
 
-        (new recalc_version_hash())->execute($logger, $stage);
+        recalc_version_hash::recalculate_version_hash($logger);
     }
 }